BRMS-36-Matching-service: updated matching service, updated test.
