+++ test/uk/gov/hmrc/brm/controllers/BirthEventsControllerSpec.scala
@@ -23,10 +23,11 @@ import play.api.libs.json.{JsValue, Json}
 import play.api.test.Helpers._
 import play.api.test.{FakeApplication, FakeRequest}
 import uk.gov.hmrc.brm.connectors.{BirthConnector, NirsConnector, NrsConnector}
-import uk.gov.hmrc.brm.services.LookupService
+import uk.gov.hmrc.brm.services.{LookupService, MatchingService}
 import uk.gov.hmrc.brm.utils.JsonUtils
 import uk.gov.hmrc.play.http.{BadRequestException, HttpResponse, NotFoundException, Upstream4xxResponse, Upstream5xxResponse}
 import uk.gov.hmrc.play.test.{UnitSpec, WithFakeApplication}
+import uk.gov.hmrc.brm.utils.TestHelper._
 
 import scala.concurrent.Future
 
@@ -75,318 +76,23 @@ class BirthEventsControllerSpec
     * */
 
   val groJsonResponseObject = JsonUtils.getJsonFromFile("500035710")
+  val groJsonResponseObject20120216 = JsonUtils.getJsonFromFile("2012-02-16")
   val groJsonResponseObject20090701 = JsonUtils.getJsonFromFile("2009-07-01")
   val groJsonResponseObject20090630 = JsonUtils.getJsonFromFile("2009-06-30")
 
-  val invalidResponse = Json.parse(
-    """
-      |[]
-    """.stripMargin)
 
-  val noJson = Json.parse(
-    s"""{
-        }
-    """.stripMargin)
-
-  val userWhereBirthRegisteredNI = Json.parse(
-    s"""
-       |{
-       | "birthReferenceNumber" : "123456789",
-       | "firstName" : "Chris",
-       | "lastName" : "Jones",
-       | "dateOfBirth" : "2012-02-16",
-       | "whereBirthRegistered" : "northern ireland"
-       |}
-    """.stripMargin)
-
-  val userWhereBirthRegisteredScotland = Json.parse(
-    s"""
-       |{
-       | "birthReferenceNumber" : "123456789",
-       | "firstName" : "Chris",
-       | "lastName" : "Jones",
-       | "dateOfBirth" : "2012-02-16",
-       | "whereBirthRegistered" : "scotland"
-       |}
-    """.stripMargin)
-
-  val userNoMatchExcludingReferenceKey = Json.parse(
-    s"""
-       |{
-       | "firstName" : "Chris",
-       | "lastName" : "Jones",
-       | "dateOfBirth" : "2012-02-16",
-       | "whereBirthRegistered" : "england"
-       |}
-    """.stripMargin)
-
-  val userNoMatchExcludingReferenceValue = Json.parse(
-    s"""
-       |{
-       | "firstName" : "Chris",
-       | "lastName" : "Jones",
-       | "dateOfBirth" : "2012-02-16",
-       | "birthReferenceNumber" : "",
-       | "whereBirthRegistered" : "england"
-       |}
-    """.stripMargin)
-
-  val userNoMatchExcludingFirstNameKey = Json.parse(
-    s"""
-       |{
-       |"lastName" : "Jones",
-       |"dateOfBirth" : "2012-04-18",
-       |"birthReferenceNumber" : "123456789"
-       |}
-     """.stripMargin)
-
-  val userNoMatchExcludingReferenceNumber = Json.parse(
-    s"""
-       |{
-       | "firstName" : "Chris",
-       | "lastName" : "Jones",
-       | "dateOfBirth" : "2012-02-16",
-       | "birthReferenceNumber" : "",
-       | "whereBirthRegistered" : "england"
-       |}
-    """.stripMargin)
-
-  val userNoMatchIncludingReferenceNumber = Json.parse(
-    s"""
-       |{
-       | "firstName" : "Chris",
-       | "lastName" : "Jones",
-       | "dateOfBirth" : "2012-08-03",
-       | "birthReferenceNumber" : "123456789",
-       | "whereBirthRegistered" : "wales"
-       |}
-    """.stripMargin)
-
-  val userNoMatchIncludingReferenceNumberCamelCase = Json.parse(
-    s"""
-       |{
-       | "firstName" : "Chris",
-       | "lastName" : "Jones",
-       | "dateOfBirth" : "2012-08-03",
-       | "birthReferenceNumber" : "123456789",
-       | "whereBirthRegistered" : "WalEs"
-       |}
-    """.stripMargin)
-
-  val userNoMatchIncludingReferenceCharacters = Json.parse(
-    s"""
-       |{
-       | "firstName" : "Chris",
-       | "lastName" : "Jones",
-       | "dateOfBirth" : "2012-08-03",
-       | "birthReferenceNumber" : "ab1_-CD263",
-       | "whereBirthRegistered" : "wales"
-       |}
-    """.stripMargin)
-
-  val userNoMatchIncludingInvalidData = Json.parse(
-    s"""
-       |{
-       | "firstName" : "Chris",
-       | "lastName" : "Jones",
-       | "dateOfBirth" : "2012-08-03",
-       | "birthReferenceNumber" : "123*34)",
-       | "whereBirthRegistered" : "wales"
-       |}
-    """.stripMargin)
-
-  val userMatchExcludingReferenceNumber = Json.parse(
-    s"""
-       |{
-       | "firstName" : "Adam TEST",
-       | "lastName" : "SMITH",
-       | "dateOfBirth" : "2012-02-16",
-       | "birthReferenceNumber" : "",
-       | "whereBirthRegistered" : "england"
-       |}
-    """.stripMargin)
-
-  val userMatchIncludingReferenceNumber = Json.parse(
-    s"""
-       |{
-       | "firstName" : "Adam TEST",
-       | "lastName" : "SMITH",
-       | "dateOfBirth" : "2012-02-16",
-       | "birthReferenceNumber" : "500035710",
-       | "whereBirthRegistered" : "england"
-       |}
-    """.stripMargin)
-
-  val userMatchCountryNameInMixCase = Json.parse(
-    s"""
-       |{
-       | "firstName" : "Adam TEST",
-       | "lastName" : "SMITH",
-       | "dateOfBirth" : "2012-02-16",
-       | "birthReferenceNumber" : "500035710",
-       | "whereBirthRegistered" : "EngLand"
-       |}
-    """.stripMargin)
-
-  val userNoMatchExcludingDateOfBirthKey = Json.parse(
-    s"""
-       |{
-       | "firstName" : "Adam TEST",
-       | "lastName" : "SMITH",
-       | "birthReferenceNumber" : "500035710",
-       | "whereBirthRegistered" : "england"
-       |}
-    """.stripMargin)
-
-  val userNoMatchExcludingDateOfBirthValue = Json.parse(
-    s"""
-       |{
-       | "firstName" : "Adam TEST",
-       | "lastName" : "SMITH",
-       | "dateOfBirth" : "",
-       | "birthReferenceNumber" : "500035710",
-       | "whereBirthRegistered" : "england"
-       |}
-    """.stripMargin)
-
-  val userNoMatchExcludingfirstNameKey = Json.parse(
-    s"""
-       |{
-       |"lastName" : "Smith",
-       |"dateOrBirth" : "2012-12-17",
-       |"birthReferenceNumber" : "123456789"
-       |}
-     """.stripMargin)
-
-
-  val userInvalidWhereBirthRegistered = Json.parse(
-    s"""
-       |{
-       |"firstname" : "Adam TEST",
-       |"lastName" : "SMITH",
-       |"dateOfBirth" : "2012-11-16",
-       |"birthReferenceNumber" : "500035710",
-       |"whereBirthRegistered": "fiji"
-       |}
-     """.stripMargin)
-
-  val userNoMatchExcludingfirstNameValue = Json.parse(
-    s"""
-       |{
-       |"firstname" : "",
-       |"lastName" : "Jones",
-       |"dateOfBirth" : "2012-11-16",
-       |"birthReferenceNumber" : "123456789"
-       |}
-     """.stripMargin)
-
-  val userNoMatchExcludinglastNameKey = Json.parse(
-    s"""
-       |{
-       |"firstname" : "John",
-       |"dateOrBirth" : "2012-12-17",
-       |"birthReferenceNumber" : "123456789"
-       |}
-     """.stripMargin)
-
-  val userNoMatchExcludinglastNameValue = Json.parse(
-    s"""
-       |{
-       |"firstname" : "John",
-       |"lastName" : "",
-       |"dateOfBirth" : "2012-11-16",
-       |"birthReferenceNumber" : "123456789"
-       |}
-     """.stripMargin)
-
-
-  val userNoMatchExcludingWhereBirthRegisteredKey = Json.parse(
-    s"""
-       |{
-       |"firstname" : "Manish",
-       |"lastName" : "Varma",
-       |"dateOfBirth" : "2012-11-16",
-       |"birthReferenceNumber" : "123456789"
-       |
-       |}
-     """.stripMargin)
-
-  val userNoMatchExcludingWhereBirthRegisteredValue = Json.parse(
-    s"""
-       |{
-       |"firstname" : "John",
-       |"lastName" : "Jones",
-       |"dateOfBirth" : "2012-11-16",
-       |"birthReferenceNumber" : "123456789",
-       |"whereBirthRegistered" : ""
-       |}
-     """.stripMargin)
-
-  val userInvalidDOB = Json.parse(
-    s"""
-       |{
-       | "firstName" : "Adam TEST",
-       | "lastName" : "SMITH",
-       | "dateOfBirth" : "2006-02-16",
-       | "birthReferenceNumber" : "500035710",
-       | "whereBirthRegistered" : "england"
-       |}
-    """.stripMargin)
-
-  val userInvalidDOBFormat = Json.parse(
-    s"""
-       |{
-       | "firstName" : "Adam TEST",
-       | "lastName" : "SMITH",
-       | "dateOfBirth" : "1234567890",
-       | "birthReferenceNumber" : "500035710",
-       | "whereBirthRegistered" : "england"
-       |}
-    """.stripMargin)
-
-  val userValidDOB = Json.parse(
-    s"""
-       |{
-       | "firstName" : "Adam TEST",
-       | "lastName" : "SMITH",
-       | "dateOfBirth" : "2012-02-16",
-       | "birthReferenceNumber" : "500035710",
-       | "whereBirthRegistered" : "england"
-       |}
-    """.stripMargin)
-
-  val userValidDOB20090701 = Json.parse(
-    s"""
-       |{
-       | "firstName" : "Adam TEST",
-       | "lastName" : "SMITH",
-       | "dateOfBirth" : "2009-07-01",
-       | "birthReferenceNumber" : "500035710",
-       | "whereBirthRegistered" : "england"
-       |}
-    """.stripMargin)
-
-  val userValidDOB20090630 = Json.parse(
-    s"""
-       |{
-       | "firstName" : "Adam TEST",
-       | "lastName" : "SMITH",
-       | "dateOfBirth" : "2009-06-30",
-       | "birthReferenceNumber" : "500035710",
-       | "whereBirthRegistered" : "england"
-       |}
-    """.stripMargin)
 
   def postRequest(v: JsValue): FakeRequest[JsValue] = FakeRequest("POST", "/api/v0/events/birth")
     .withHeaders((ACCEPT, "application/vnd.hmrc.1.0+json"), ("Audit-Source", "DFS"))
     .withBody(v)
 
   val mockConnector = mock[BirthConnector]
-
+  val mockMatchingService = mock[MatchingService]
   object MockLookupService extends LookupService {
     override val groConnector = mockConnector
     override val nirsConnector = NirsConnector
     override val nrsConnector = NrsConnector
+    override val matchingService = MatchingService
   }
 
   object MockController extends BirthEventsController {
@@ -731,7 +437,7 @@ class BirthEventsControllerSpec
 
       "return matched value of true when the dateOfBirth is greater than 2009-07-01 and the gro record matches" in {
         running(FakeApplication(additionalConfiguration = config)) {
-          when(MockController.service.groConnector.getReference(Matchers.any())(Matchers.any())).thenReturn(Future.successful(httpResponse(groJsonResponseObject)))
+          when(MockController.service.groConnector.getReference(Matchers.any())(Matchers.any())).thenReturn(Future.successful(httpResponse(groJsonResponseObject20120216)))
           val request = postRequest(userValidDOB)
           val result = MockController.post().apply(request)
           status(result) shouldBe OK
